requirements.txt
django>=2.0,<2.1
djangorestframework>=3.8.2,<3.9
markdown>=2.6.11,<2.7
django-filter>=1.1,<1.2
https://github.com/nezhar/django-rest-multiauthtoken/archive/master.zip